+++ config/chief-settings.php
@@ -3,17 +3,6 @@
 return [
 
     /**
-     * Homepage id
-     *
-     * Here you can explicitly set the page that is considered to be the website homepage,
-     * This page is always retrieved via the root url and does not have an own url.
-     * If value is left empty, the first published page is used as a default.
-     *
-     * e.g. 'homepage' => 2,
-     */
-    'homepage' => null,
-
-    /**
      * Contact person (aka webmaster)
      *
      * The contact person receives all incoming communication e.g. contact form submissions