+++ resources/views/back/managers/edit.blade.php
@@ -30,7 +30,10 @@
 
         <!-- needs to be before form to be detected by context-menu. Don't know why :s -->
         @include('chief::back.managers._partials.delete-modal')
-        @include('chief::back.managers._partials.archive-modal')
+
+        @if($manager->isAssistedBy('archive'))
+            @include('chief::back.managers._partials.archive-modal')
+        @endif
 
         <form id="updateForm" method="POST" action="{{ $manager->route('update') }}" enctype="multipart/form-data" role="form">
             {{ csrf_field() }}