+++ src/Pages/NotFoundHomepage.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Thinktomorrow\Chief\Pages;
-
-class NotFoundHomepage extends \Exception
-{
-}