+++ src/Urls/ChiefResponse.php
@@ -12,17 +12,17 @@ use Thinktomorrow\Chief\Concerns\Publishable\PreviewMode;
 
 class ChiefResponse extends Response
 {
-    public static function fromRequest(Request $request = null, $locale = null)
-    {
-        if (!$request) {
-            $request = request();
-        }
-        if (!$locale) {
-            $locale = app()->getLocale();
-        }
-
-        return static::fromSlug($request->path(), $locale);
-    }
+//    public static function fromRequest(Request $request = null, $locale = null)
+//    {
+//        if (!$request) {
+//            $request = request();
+//        }
+//        if (!$locale) {
+//            $locale = app()->getLocale();
+//        }
+//
+//        return static::fromSlug($request->path(), $locale);
+//    }
 
     public static function fromSlug(string $slug, $locale = null)
     {
@@ -48,8 +48,6 @@ class ChiefResponse extends Response
 
                 return static::createRedirect($model->url($locale));
             }
-
-
 
             if (method_exists($model, 'isPublished') && ! $model->isPublished()) {